requirements.txt
pyyaml
readchar
playsound3